- Moved GameDebugTheme enum into correct file - Fixes after rebase
