+++ src/components/ui/debug/GameDebugToggleButton.tsx
@@ -5,7 +5,7 @@ import {useKeyboardInput} from "../../../utils/hooks";
 import {Key} from "ts-key-enum";
 import gameDebugStore from "../../../store/GameDebugStore";
 import DebugButton from "./DebugButton";
-import {GameDebugTheme} from "../../../utils/interfaces";
+import {GameDebugTheme} from "../../../utils/enums";
 
 const GameDebugToggleButton: FunctionComponent = () => {
   const {debug, showDebug, hideDebug} = gameDebugStore
