+++ src/utils/enums.ts
@@ -14,3 +14,8 @@ export enum GameState {
 export enum GameEvents {
   SpamDestroyed = 'SPAM_DESTROYED'
 }
+
+export enum GameDebugTheme {
+  Light = "LIGHT",
+  Dark = "DARK"
+}
